Make the thought cache's flushQueue method into a standalone action creator.
