+++ src/action-creators/index.ts
@@ -16,6 +16,7 @@ export { default as logout } from './logout'
 export { default as modalComplete } from './modalComplete'
 export { default as newThought } from './newThought'
 export { default as preloadSources } from './preloadSources'
+export { default as pull } from './pull'
 export { default as restoreCursorBeforeSearch } from './restoreCursorBeforeSearch'
 export { default as scrollCursorIntoView } from './scrollCursorIntoView'
 export { default as setEditingValue } from './setEditingValue'
