+++ src/action-creators/pull.ts
@@ -1,0 +1,111 @@
+import _ from 'lodash'
+import * as db from '../data-providers/dexie'
+import * as firebaseProvider from '../data-providers/firebase'
+import getManyDescendants from '../data-providers/data-helpers/getManyDescendants'
+import { mergeThoughts } from '../util'
+import { ActionCreator, Context, Index, Lexeme, Parent, ThoughtsInterface } from '../types'
+
+const BUFFER_DEPTH = 2
+
+export interface PullOptions {
+  maxDepth?: number,
+}
+
+/** Iterate through an async iterable and invoke a callback on each yield. */
+async function itForEach<T> (it: AsyncIterable<T>, callback: (value: T) => void) {
+  // eslint-disable-next-line fp/no-loops
+  for await (const item of it) {
+    callback(item)
+  }
+}
+
+/**
+ * Fetch descendants of thoughts.
+ * WARNING: Unknown behavior if thoughtsPending takes longer than throttleFlushPending.
+ */
+const pull = (pending: Index<Context>, { maxDepth }: PullOptions = {}): ActionCreator => async (dispatch, getState) => {
+
+  if (Object.keys(pending).length === 0) return
+
+  // get local thoughts
+  const thoughtLocalChunks: ThoughtsInterface[] = []
+
+  const thoughtsLocalIterable = getManyDescendants(db, pending, { maxDepth: maxDepth || BUFFER_DEPTH })
+  for await (const thoughts of thoughtsLocalIterable) { // eslint-disable-line fp/no-loops
+
+    // eslint-disable-next-line fp/no-mutating-methods
+    thoughtLocalChunks.push(thoughts)
+
+    // TODO: Update only thoughts for which shouldUpdate is false in reconcile and remove redundant updateThoughts. Entries for which shouldUpdate is true are updated anyway.
+    // mergeUpdates will prevent overwriting non-pending thoughts with pending thoughts
+    dispatch({
+      type: 'updateThoughts',
+      contextIndexUpdates: thoughts.contextIndex,
+      thoughtIndexUpdates: thoughts.thoughtIndex,
+      local: false,
+      remote: false,
+    })
+  }
+
+  const thoughtsLocal = thoughtLocalChunks.reduce(_.ary(mergeThoughts, 2))
+
+  // get remote thoughts and reconcile with local
+  const user = getState().user
+  if (user) {
+
+    const thoughtsRemoteIterable = getManyDescendants(firebaseProvider, pending, { maxDepth: maxDepth || BUFFER_DEPTH })
+
+    const thoughtRemoteChunks: ThoughtsInterface[] = []
+
+    // TODO: Refactor into zipThoughts
+    await itForEach(thoughtsRemoteIterable, (thoughtsRemoteChunk: ThoughtsInterface) => {
+
+      // eslint-disable-next-line fp/no-mutating-methods
+      thoughtRemoteChunks.push(thoughtsRemoteChunk)
+
+      // find the corresponding Parents from the local store (if any exist) so it can be reconciled with the remote Parents
+      const thoughtsLocalContextIndexChunk = _.transform(thoughtsRemoteChunk.contextIndex, (accum, parentEntryRemote, key) => {
+        const parentEntryLocal = thoughtsLocal.contextIndex[key]
+        if (parentEntryLocal) {
+          accum[key] = parentEntryLocal
+        }
+      }, {} as Index<Parent>)
+
+      // find the corresponding Lexemes from the local store (if any exist) so it can be reconciled with the remote Lexemes
+      const thoughtsLocalThoughtIndexChunk = _.transform(thoughtsRemoteChunk.thoughtIndex, (accum, lexemeRemote, key) => {
+        const lexemeLocal = thoughtsLocal.thoughtIndex[key]
+        if (lexemeLocal) {
+          accum[key] = lexemeLocal
+        }
+      }, {} as Index<Lexeme>)
+
+      dispatch({
+        type: 'reconcile',
+        thoughtsResults: [{
+          contextIndex: thoughtsLocalContextIndexChunk,
+          thoughtIndex: thoughtsLocalThoughtIndexChunk,
+        }, thoughtsRemoteChunk]
+      })
+
+    })
+
+    const thoughtsRemote = thoughtRemoteChunks.reduce(_.ary(mergeThoughts, 2))
+
+    // the reconcile dispatched above is based on remote keys only
+    // thoughts that exist locally and not remotely will be missed
+    // sync all thoughts here to ensure none are missed
+    // TODO: Only reconcile local-only thoughts
+    dispatch({
+      type: 'reconcile',
+      thoughtsResults: [thoughtsLocal, thoughtsRemote]
+    })
+  }
+
+  // If the buffer size is reached on any loaded thoughts that are still within view, we will need to invoke flushPending recursively. Queueing updatePending will properly check visibleContexts and fetch any pending thoughts that are visible.
+  const hasPending = Object.keys(thoughtsLocal.contextIndex || {})
+    .some(key => (thoughtsLocal.contextIndex || {})[key].pending)
+
+  return hasPending
+}
+
+export default pull
